requirements.txt
Django>=4.0
djangorestframework
psycopg2-binary
python-dotenv
drf-spectacular
drf-spectacular-sidecar
django-cors-headers>=4.0.0
djangorestframework-simplejwt
drf_standardized_errors
google-auth
requests